many fixes, inluding: 1) shared memory usage is oracle; 2) "drop pixel problem" fix (with edge resize, without double rendering); 3) CSAA + MSAA support; And other small fixes;

+++ src/DecisionBits.h
@@ -2,7 +2,7 @@
 #include "bothInclude.h"
 
 struct DecisionBits {
-    uint16_t arr;
+    int arr;
 
     __host__ __device__
     DecisionBits() { }
